requirements.txt
opencv-python==4.6.0.66
locust==2.10.1
python-dotenv==0.20.0